# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Clinic management API. Every request runs inside the clinic bound to the caller's session,\nand reads never cross that boundary.\n\nAuthentication is a session cookie: `POST /api/session` opens it, `DELETE /api/session` closes it,\nand every other operation requires it.\n\nEndpoints belonging to an optional module answer `404` while the clinic has not activated that\nmodule, so a client should treat their absence as a feature the clinic did not contract.\n`GET /api/capabilities` reports what is currently available.\n\nErrors always carry the same body: status, message, request path and, for a rejected payload,\nthe offending fields.",
    "title" : "Clivo API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "sessionCookie" : [ ]
  } ],
...
  "paths" : {
...
    "/api/notifications" : {
      "get" : {
        "operationId" : "listAppointmentNotifications",
        "parameters" : [ {
          "in" : "query",
          "name" : "appointmentId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/NotificationView"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "List an appointment's notifications",
        "tags" : [ "Notifications" ]
      },
      "post" : {
        "operationId" : "scheduleNotification",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NotificationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NotificationView"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Schedule a reminder for an appointment",
        "tags" : [ "Notifications" ]
      }
    },
    "/api/notifications/dispatch" : {
      "post" : {
        "operationId" : "dispatchDueNotifications",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/NotificationView"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Dispatch every notification already due",
        "tags" : [ "Notifications" ]
      }
    },
    "/api/notifications/pending" : {
      "get" : {
        "operationId" : "listPendingNotifications",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/NotificationView"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "List the notifications not dispatched yet",
        "tags" : [ "Notifications" ]
      }
    },
    "/api/notifications/{id}/reply" : {
      "post" : {
        "operationId" : "registerNotificationReply",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReplyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/NotificationView"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Register the customer's reply to a notification",
        "tags" : [ "Notifications" ]
      }
    },
...
    "/api/session-packages" : {
      "get" : {
        "operationId" : "listCustomerSessionPackages",
        "parameters" : [ {
          "in" : "query",
          "name" : "customerId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PackageView"
                  }
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "List a customer's packages",
        "tags" : [ "Session packages" ]
      },
      "post" : {
        "operationId" : "sellSessionPackage",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PackageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PackageView"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Sell a package of sessions to a customer",
        "tags" : [ "Session packages" ]
      }
    },
    "/api/session-packages/{id}" : {
      "get" : {
        "operationId" : "getSessionPackage",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PackageView"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Read one package with the sessions already used",
        "tags" : [ "Session packages" ]
      }
    },
    "/api/session-packages/{id}/cancellation" : {
      "post" : {
        "operationId" : "cancelSessionPackage",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PackageView"
                }
              }
            },
            "description" : "OK"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not Found"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unprocessable Content"
          }
        },
        "summary" : "Cancel a package, releasing the sessions not used",
        "tags" : [ "Session packages" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "violations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            }
          }
        }
      },
...
      "FieldViolation" : {
        "type" : "object",
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "NotificationRequest" : {
        "type" : "object",
        "properties" : {
          "appointmentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "channel" : {
            "type" : "string",
            "minLength" : 1
          },
          "recipient" : {
            "type" : "string"
          }
        },
        "required" : [ "appointmentId", "channel" ]
      },
      "NotificationView" : {
        "type" : "object",
        "properties" : {
          "appointmentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "appointmentStart" : {
            "type" : "string",
            "format" : "date-time"
          },
          "channel" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "recipient" : {
            "type" : "string"
          },
          "reply" : {
            "type" : "string"
          },
          "sentAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "PackageRequest" : {
        "type" : "object",
        "properties" : {
          "customerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expiresOn" : {
            "type" : "string",
            "format" : "date"
          },
          "price" : {
            "type" : "number"
          },
          "serviceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalSessions" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "required" : [ "customerId", "price", "serviceId" ]
      },
      "PackageView" : {
        "type" : "object",
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "customerId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expiresOn" : {
            "type" : "string",
            "format" : "date"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "price" : {
            "type" : "number"
          },
          "remainingSessions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "serviceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "serviceName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "totalSessions" : {
            "type" : "integer",
            "format" : "int32"
          },
          "usedSessions" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ReplyRequest" : {
        "type" : "object",
        "properties" : {
          "reply" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "sessionCookie" : {
        "description" : "Session cookie issued by POST /api/session.",
        "in" : "cookie",
        "name" : "JSESSIONID",
        "type" : "apiKey"
      }
    }
  }
}